+++ lib/screen/footPrint/footprint_date_plan_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class FootprintDatePlanScreen extends StatefulWidget {
+  const FootprintDatePlanScreen({super.key});
+
+  @override
+  State<FootprintDatePlanScreen> createState() => _FootprintDatePlanScreenState();
+}
+
+class _FootprintDatePlanScreenState extends State<FootprintDatePlanScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text("dateplan"));
+  }
+}
