+++ pubspec.yaml
@@ -36,12 +36,18 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
   provider: ^6.1.2
-  flutter_svg: ^2.0.10+1
+  flutter_svg: ^2.0.9
   flutter_staggered_grid_view: ^0.7.0
   table_calendar: ^3.1.1
   intl: ^0.19.0
   image_picker: ^1.1.1
   fluttertoast: ^8.2.5
+  flutter_map: ^4.0.0
+  latlong2: ^0.8.2
+  geojson: ^1.0.0
+  flutter_carousel_widget: ^2.2.0
+
+
 
 dev_dependencies:
   flutter_test:
@@ -69,6 +75,7 @@ flutter:
   assets:
     - lib/assets/icons/
     - lib/assets/images/
+    - lib/assets/geojsons/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
